+++ pkg/dynamic/viper.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io"
 
-	"github.com/dgrijalva/jwt-go"
+	"github.com/golang-jwt/jwt/v4"
 
 	"github.com/fsnotify/fsnotify"
 	"github.com/spf13/viper"