+++ pkg/jwt/token_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"time"
 
-	jwtgo "github.com/dgrijalva/jwt-go"
 	"github.com/gofrs/uuid"
+	jwtgo "github.com/golang-jwt/jwt/v4"
 	"github.com/stretchr/testify/assert"
 )