+++ pkg/middlewares/auth_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/gesundheitscloud/go-svc/pkg/instrumented"
 
-	"github.com/dgrijalva/jwt-go"
+	"github.com/golang-jwt/jwt/v4"
 
 	"github.com/stretchr/testify/assert"
 )