+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+- Replace vulnerable `dgrijalva/jwt-go` with `golang-jwt/jwt`
+
 ## [v1.14.1] - 2021-08-05
 
 ### Fixed