+++ CHANGELOG.md
@@ -11,8 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Inconsistent BI events names: changed `token-refreshed` to `token-refresh`, `sharing-revoked` to `sharing-revoke`
-
 ### Deprecated
 
 ### Removed
@@ -20,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [v1.17.0] - 2021-08-23
+
+### Changed
+
+- Inconsistent BI events names: changed `token-refreshed` to `token-refresh`, `sharing-revoked` to `sharing-revoke`
 
 ## [v1.16.0] - 2021-08-13
 
@@ -380,7 +384,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial state: standards for Main, HTTP Server, DB access (gorm), Logging, Instrumented-Handler, and K8s Probe
 
-[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.16.0...HEAD
+[Unreleased]: https://github.com/gesundheitscloud/go-svc/compare/v1.17.0...HEAD
+[v1.17.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.16.0...v1.17.0
 [v1.16.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.15.0...v1.16.0
 [v1.15.0]: https://github.com/gesundheitscloud/go-svc/compare/v1.14.1...v1.15.0
 [v1.14.1]: https://github.com/gesundheitscloud/go-svc/compare/v1.14.0...v1.14.1